Removed joins code
